requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.5.14
docx2txt==0.8
emoji==2.14.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
Faker==30.3.0
fastapi
favicon==0.7.0
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
htbuilder==0.6.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.0.201
langchainplus-sdk==0.0.20
langdetect==1.0.9
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.3
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numexpr==2.8.6
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.5
packaging==23.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==4.25.5
pyarrow==17.0.0
pycparser==2.22
pydantic
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.11.2
pyparsing==3.1.4
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.4.27
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.12
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
st-annotated-text==4.0.1
starlette==0.39.2
streamlit==1.27.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
unstructured
unstructured-client
urllib3==2.2.3
uvicorn==0.31.1
validators==0.34.0
watchdog==4.0.2
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.15.2
zipp==3.20.2
